Fix UTC Date handling
When selecting a date, if the user's timezone was less that UTC, the
created date object always rolled over to the previous day from the
UTC date created from the selected year/month/day values.

The solution was to check the UTC year/month/day of the day object
against the selected values, not the year/month/day of the local
timezone.